+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="2.3.9"></a>
 # 2.3.9 (08.08.2018)
 
+### Feature
+* **terra-breadcrumbs** added. Breadcrumbs using router and routerLink.
+
 ### Bug Fixes
 * **terra-suggestion-box** reset selected value when `inputListBoxValues` is empty
 * **style** fixed scss variables